create EventCard component
